+++ internal/acceptance_tests/blocks/logging_kinesis_nested.tf
@@ -1,0 +1,9 @@
+logging_kinesis {
+  name  = "{{.LOGGING_KINESIS_NAME}}"
+  topic = "test-stream"
+
+  authentication = {
+    access_key = "AKIAIOSFODNN7EXAMPLE"
+    secret_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
+  }
+}