+++ internal/acceptance_tests/blocks/logging_kinesis_nested_multi.tf
@@ -1,0 +1,19 @@
+logging_kinesis {
+  name  = "{{.LOGGING_KINESIS_NAME_1}}"
+  topic = "test-stream"
+
+  authentication = {
+    access_key = "AKIAIOSFODNN7EXAMPLE"
+    secret_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
+  }
+}
+
+logging_kinesis {
+  name  = "{{.LOGGING_KINESIS_NAME_2}}"
+  topic = "test-stream"
+
+  authentication = {
+    access_key = "AKIAIOSFODNN7EXAMPLE"
+    secret_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY"
+  }
+}